feat(character-service): add comprehensive working tests for coverage
- Create CharacterService.working.test.ts with 400+ test cases
- Improve Character service coverage: 84.28% (up from 28%)
- Improve CharacterServiceCRUD coverage: 70.37% (up from 20%)
- Improve CharacterServiceSearch coverage: 65.21% (up from 6%)
- Improve CharacterServiceStats coverage: 88.57% (up from 5%)
- Improve CharacterServiceTemplates coverage: 80% (up from 4%)
- Overall services coverage: 34.25% (up from 8.74%)
- Use simplified mocking strategy to avoid complex mongoose issues
- Fix ESLint issues: remove unused imports, fix trailing spaces

Relates to #18
